Add files via upload
The ATM App provides the user with 2 options for withdrawal: QR and SMS. The QR code will be displayed and if user opts for withdrawal through QR, the scanner on the user app must scan the QR being displayed on the ATM app. The app is then redirected to a page where the amount to be withdrawn is shown and can be edited if the user wants to. He should then click on Withdraw Cash button. Then, the remaining balance is displayed after withdrawal of cash from the balance. 
If the user wants to withdraw through SMS, one must enter phone no. and click on Verify button. One must then enter the otp sent to one's mobile. The app then redirects to cash withdrawal page where money can be edited and withdrawn.